source-systems/data-generation/requirements.txt
pandas>=2.0.0
faker>=20.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
boto3>=1.28.0
pyarrow>=12.0.0
python-dotenv>=1.0.0